requirements.txt
pycognito==2024.5.1
requests>=2.31.0
python-dotenv>=1.0.0
